pyproject.toml
[tool.poetry]
name = "open-geo-engine"
version = "0.1.0"
description = ""
authors = ["ChristinaLast <christina.last@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^21.7b0"
pylama = "^7.7.1"
bandit = "^1.7.0"
Django = "^3.2.6"
PyYAML = "^5.4.1"
pydantic = "^1.8.2"
loguru = "^0.5.3"
joblib = "^1.0.1"
libpysal = "^4.5.1"
geopandas = "^0.9.0"
esda = "^2.4.1"
sklearn = "^0.0"
catboost = "^0.26.1"
mlflow = "^1.20.1"
click-config-file = "^0.6.0"
click-option-group = "^0.5.3"
hydra-core = "^1.1.1"
earthengine-api = "^0.1.280"
setuptools = "^57.4.0"
osmnx = "^1.1.1"
google = "^3.0.0"
google-streetview = "^1.2.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyterlab = "^3.1.9"
ipykernel = "^6.2.0"
mypy = "^0.910"
coverage = "^5.5"
pre-commit = "^2.14.0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
